wx_objects: Fix calls to unexported gen_server:print_event/3
In code copied (and modified) from the gen_server module, there
are references to the gen_server:print_event/3 function, which is
not exported.

We could export print_event/3 from the gen_server, but in that
case we must promise to keep the function and keep it compatible
in the future. Since the function is small and wx_object already
contains a lot of code copied from gen_server, let's copy the
code for print_event/3 too.

(Thanks to Kostis Sagonas. This old bug was exposed by a patch
by Kostis that replaced "tuple funs" with real funs.)